* [Platform] All DAO::delete() calls will now properly bailout if the ID list they're given is empty.  Normally this is not necessary, but there have been a few situations where users have cached URLs in their local browser that attempt to repeat actions (like deleting worklists or tasks).

+++ api/DAO.class.php
@@ -1108,6 +1108,9 @@ class DAO_WorkerEvent extends DevblocksORMHelper {
 		if(!is_array($ids)) $ids = array($ids);
 		$db = DevblocksPlatform::getDatabaseService();
 		
+		if(empty($ids))
+			return;
+		
 		$ids_list = implode(',', $ids);
 		
 		$db->Execute(sprintf("DELETE FROM worker_event WHERE id IN (%s)", $ids_list));
@@ -1302,6 +1305,9 @@ class DAO_ContactOrg extends C4_ORMHelper {
 	static function delete($ids) {
 		if(!is_array($ids)) $ids = array($ids);
 		$db = DevblocksPlatform::getDatabaseService();
+		
+		if(empty($ids))
+			return;
 		
 		$id_list = implode(',', $ids);
 		
@@ -1669,7 +1675,9 @@ class DAO_Address extends C4_ORMHelper {
 	
     static function delete($ids) {
         if(!is_array($ids)) $ids = array($ids);
-        if(empty($ids)) return;
+
+		if(empty($ids))
+			return;
 
 		$db = DevblocksPlatform::getDatabaseService();
         
@@ -2376,7 +2384,9 @@ class DAO_MessageNote extends DevblocksORMHelper {
     
     static function delete($ids) {
         if(!is_array($ids)) $ids = array($ids);
-        if(empty($ids)) return;
+
+		if(empty($ids))
+			return;
 
 		$db = DevblocksPlatform::getDatabaseService();
         
@@ -4024,7 +4034,9 @@ class DAO_ViewRss extends DevblocksORMHelper {
 	}
 	
 	static function delete($id) {
-		if(empty($id)) return;
+		if(empty($id))
+			return;
+		
 		$db = DevblocksPlatform::getDatabaseService();
 		
 		$sql = sprintf("DELETE QUICK FROM view_rss WHERE id = %d",
@@ -4630,6 +4642,9 @@ class DAO_Bucket extends DevblocksORMHelper {
 	    if(!is_array($ids)) $ids = array($ids);
 		$db = DevblocksPlatform::getDatabaseService();
 		
+		if(empty($ids))
+			return;
+		
 		/*
 		 * Notify anything that wants to know when buckets delete.
 		 */
@@ -4852,6 +4867,10 @@ class DAO_MailToGroupRule extends DevblocksORMHelper {
 	
 	static function delete($ids) {
 		if(!is_array($ids)) $ids = array($ids);
+		
+		if(empty($ids))
+			return;
+		
 		$db = DevblocksPlatform::getDatabaseService();
 		
 		$ids_list = implode(',', $ids);
@@ -4953,6 +4972,9 @@ class DAO_Community extends DevblocksORMHelper {
 	    if(!is_array($ids)) $ids = array($ids);
 	    $db = DevblocksPlatform::getDatabaseService();
 	    
+		if(empty($ids))
+			return;
+		
 	    $id_list = implode(',', $ids);
 	    
 	    $sql = sprintf("DELETE QUICK FROM community WHERE id IN (%s)", $id_list);
@@ -5164,6 +5186,9 @@ class DAO_WorkerWorkspaceList extends DevblocksORMHelper {
 	
 	static function delete($ids) {
 		if(!is_array($ids)) $ids = array($ids);
+		
+		if(empty($ids))
+			return;
 		
 		$db = DevblocksPlatform::getDatabaseService();
 		$ids_list = implode(',', $ids);
@@ -5401,6 +5426,10 @@ class DAO_Note extends DevblocksORMHelper {
     
 	static function delete($ids) {
 		if(!is_array($ids)) $ids = array($ids);
+		
+		if(empty($ids))
+			return;
+		
 		$db = DevblocksPlatform::getDatabaseService();
 		
 		$ids_list = implode(',', $ids);
@@ -5558,6 +5587,9 @@ class DAO_PreParseRule extends DevblocksORMHelper {
 		if(!is_array($ids)) $ids = array($ids);
 		$db = DevblocksPlatform::getDatabaseService();
 		
+		if(empty($ids))
+			return;
+		
 		$ids_list = implode(',', $ids);
 		
 		$db->Execute(sprintf("DELETE QUICK FROM preparse_rule WHERE id IN (%s)", $ids_list));
@@ -5709,6 +5741,9 @@ class DAO_GroupInboxFilter extends DevblocksORMHelper {
 	    if(!is_array($ids)) $ids = array($ids);
 	    $db = DevblocksPlatform::getDatabaseService();
 	    
+		if(empty($ids))
+			return;
+		
 	    $id_list = implode(',', $ids);
 	    
 	    $sql = sprintf("DELETE QUICK FROM group_inbox_filter WHERE id IN (%s)", $id_list);
@@ -5875,6 +5910,9 @@ class DAO_MailTemplate extends DevblocksORMHelper {
 	public static function delete($ids) {
 		if(!is_array($ids)) $ids = array($ids);
 		$db = DevblocksPlatform::getDatabaseService();
+		
+		if(empty($ids))
+			return;
 		
 		$sql = sprintf("DELETE QUICK FROM %s WHERE id IN (%s)",
 			self::_TABLE,
@@ -6064,6 +6102,9 @@ class DAO_TicketComment extends DevblocksORMHelper {
 		if(!is_array($ids)) $ids = array($ids);
 		$db = DevblocksPlatform::getDatabaseService();
 		
+		if(empty($ids))
+			return;
+		
 		$ids_list = implode(',', $ids);
 		
 		$db->Execute(sprintf("DELETE QUICK FROM ticket_comment WHERE id IN (%s)", $ids_list));
@@ -6188,6 +6229,9 @@ class DAO_CustomField extends DevblocksORMHelper {
 	
 	public static function delete($ids) {
 		if(!is_array($ids)) $ids = array($ids);
+		
+		if(empty($ids))
+			return;
 		
 		$db = DevblocksPlatform::getDatabaseService();
 		
@@ -6808,6 +6852,9 @@ class DAO_Task extends C4_ORMHelper {
 	static function delete($ids) {
 		if(!is_array($ids)) $ids = array($ids);
 		$db = DevblocksPlatform::getDatabaseService();
+		
+		if(empty($ids))
+			return;
 		
 		$ids_list = implode(',', $ids);
